refactor: relocate content components to InfoPanel directory and update imports in TabbedPanel
