+++ src/app/components/InfoPanel/AboutContent.tsx
